cv_service/requirements.txt
# Core dependencies
flask>=3.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Computer Vision dependencies - Install these manually if auto-install fails
opencv-python-headless>=4.8.1  # Alternative to opencv-python
numpy>=1.26.0

# ML/DL dependencies - Install these manually if auto-install fails
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision

# Optional dependencies - Install these manually if needed
# mediapipe>=0.10.8
# ultralytics>=8.0.196

# Database
# Core dependencies
flask>=3.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Computer Vision dependencies
opencv-python-headless>=4.8.1
numpy>=1.26.0

# Deep Learning dependencies
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision

# Database
supabase>=1.0.3

# Additional dependencies will be handled in code with fallbacks